(chore) renaming the opa command to operator-permission-advisor so we do not conflict with the opm policy agent project
