+++ cmd/operatorPermissionAdvisor/lookupPermissions.go
